完善获取二维码以及删除二维码接口

+++ mdc-comvita-service/src/main/java/com/aquilaflycloud/mdc/result/pre/PreActivityDetailResult.java
@@ -36,6 +36,6 @@ public class PreActivityDetailResult extends PreActivityInfo {
     private List<RefShopInfoResult> refShops;
 
     @ApiModelProperty(value = "参加人数")
-    private Long participationCount;
+    private int participationCount;
 
 }
